code refactor for datasource package
